V4: Change CallValue to have the value on the stack
We used to store the value-to-be-called in the accumulator. So the
generated bytecode looked like:
    LoadReg r1
    CallValue()

The first thing done in CallValue is to store the accumulator. So by not
loading the accumulator, we can actually remove the subsequent store,
which results in less interpreter instructions and one less store in
CallValue.

Change-Id: Icc7c8a5449bf369b9226d66bc6055cb705ef660e
Reviewed-by: Lars Knoll <lars.knoll@qt.io>