Fix a bug with calculation of translation in DragHandler
The problem was that the start position of the target item
(m_targetStartPos) was not always updated.

In combination with several handlers (this bug was observed with
PinchHandler and DragHandler as siblings) it is quite likely that the
DragHandler will get a press, but not activate (because the PinchHandler
might activate instead). It would therefore not deactivate (so it would
not reset m_targetStartPos), and as a consequence the m_targetStartPos was
not updated to its correct value the next time we initiated a drag because
m_targetStartPos wasn't null.

Change-Id: Icf3089fc685cfbcbcfba638966a2e8ad1c8c089f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>