V4 JIT: Try not to overdo alignment
The x86_64 assembler will not push a register aligned. The arm64
assembler will. The correct use for pushing constants is not to
"prepare" the stack pointer and then push (because that would result in
an unaligned value on arm64), but to load the value and do an aligned
register push.

This fixes a Bus Error on ARM64.

Change-Id: I9c9ed643372025f5f7097e7f44ab17adfd0815a9
Reviewed-by: Lars Knoll <lars.knoll@qt.io>