Fix QQmlImportsPrivate::resolvedUri()
Qt Quick Controls 2 styles have a version in the middle of the path/URI
(e.g. qml/QtQuick/Controls.2/Material). The resolvedUri() helper method,
which is used by addFileImport() for implicit imports, was never taught
to deal with this type of versioning. It was only attempting to remove
the version from the end.

Change-Id: Ibdf23dc6c3b0794527d5f9330602858291c23e01
Task-number: QTBUG-64868
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>