Add optional timing logs for Shape triangulation
Follow the scenegraph's approach with qt.scenegraph.time.* categories
and add qt.shape.time.sync. When enabled, this will print a line for
each sync() that had something to process. For asynchronous items this
includes all the time from requesting the backend to start the
work right up to the finished callback, and thus can be somewhat
misleading.

Task-number: QTBUG-64951
Change-Id: Ib51a9d4adba4cdc7d692eb7ff14627a5d0e548cf
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>