TapHandler: emit canceled when a gesturePolicy constrait is violated
If gesturePolicy is DragThreshold and the threshold is exceeded or
if gesturePolicy is withinBounds and you drag outside the bounds,
that means any pending tap gesture is canceled, so the canceled
signal should be emitted.  It should not depend on a grab or an event
to be canceled.  QQuickTapHandler::setPressed(press, cancel, point)
should take care of it when cancel is true.

Task-number: QTBUG-65003
Change-Id: I2018695f5fc4438e68df3b34cf258251dbd5f198
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>