Fix custom scenegraph item animations when used with QQuickWidget
When used within a QQuickWidget, our custom scenegraph item animations
wouldn't update. Gunnar suggested that the cause of this is the fact
that we're not calling QQuickWindow::update() anywhere (as would be
done for us if we implemented these controls with Animators in QML).

To go into more detail, QQuickWidget uses a QQuickRenderControl
internally, and the render control's renderRequested() and
sceneChanged() signals are both connected to
QQuickWidget::triggerUpdate(), so if calling QQuickWindow::update()
is enough to kick it into action.

QQuickWindow::frameSwapped is not emitted as part of
QQuickWidget's render cycle, so that is probably the core of the
problem. So instead, of using frameSwapped to drive the animation,
we drive it directly from our beforeRendering handler.

We also want to ensure that the very first animation triggers a
start as well, so we make QQuickAnimatedNode::start() call
QQuickWindow::update().

It was also suggested that we use direct connections when connecting to
QQuickWindow::beforeRendering and QQuickWindow::frameSwapped, as
QQuickWindow has affinity to the GUI thread. So, to make sure that
things get called on the right thread at the right time, we change
that as well.

Task-number: QTBUG-62874
Change-Id: Id233be1b80858ec19eef9312f143adffadababff
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>