Tumbler: ensure that currentIndex animations aren't too slow
The default (for ListView) is 400 pixels a second, which is really slow
when there are 1000 items. Set highlightMoveDuration to 1000 (one
second) for both view types to ensure a constant amount of time to
animate currentIndex changes.

[ChangeLog][Controls][Tumbler] Made currentIndex animations
take a constant amount of time (1 second) regardless of how many
items are in the model. This prevents Tumblers with large amounts
of items from scrolling too slowly when changing the currentIndex.

Task-number: QTBUG-65165
Change-Id: I0342236e49fae346926eb7e8a5fb6ed0813e457e
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>