Make tst_customization easier to maintain
Maintain the tested controls and delegates in a single array that can
be re-used for testing all styles without repetitive and error-prone
hand-written data rows.

For this to work, the structure of the tested styles must match. The
"incomplete" and "override" styles filled in all possible delegates,
whereas the "simple" style filled in only the normally used delegates.
The test styles have now been synced with the structure of our built-
in styles. For example, Dials and Sliders don't normally have a content
item, and CheckBoxes and RadioButtons don't normally have a background.

Change-Id: I48a26ee170f66882c55b54a282f2e4b3a3875f9a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>