Default: eliminate internal IDs in DelayButton
An ID in the internal Text element prevented deferred execution for
the whole content item. The ID was used for two reasons. First of all,
it was used to propagate implicit size from the Text element to the
root of the content item. Secondly, it was used to calculate clip areas
to provide the same text in two different colors. This patch provides
two internal C++ helpers, ItemGroup and ClippedText, that provide these
functionalities without the need of using IDs in QML. At the same time
we got rid of two wrapper Items and simplified some QML bindings, which
results to a nice boost (18->22) in qmlbench on TX1.

Task-number: QTBUG-65341
Change-Id: Icf9c09356cf5c0ed641bde537bee7291bd260057
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>