Make AbstractButton's text win over action's when both are specified
The idea is that you share a generic Action in different places in
the UI, and then you override things locally in specific controls.
This patch enforces that by changing QQuickAbstractButton::text()
to return the Button's text if explicitly set (or there isn't an
Action), or the action's text if one is set.

Change-Id: Iae422abca35c9598f73ba7a1d3f1387c36037ec0
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>