Material ToolButton: fix round ripple for square buttons
The condition for checking whether a tool button is square and thus,
whether it should use a round ripple instead of a fill ripple, must be
based on the size of the contents, not on the implicit size. This way,
you get a round ripple for one-letter tool buttons but only if they
are not stretched.

Task-number: QTBUG-65052
Change-Id: I0d498dc65650f3b786b41e538748549d61f2b216
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>