Add creation benchmarks for qmlbench
The qmlbench repo has benchmarks only for a subset of controls. They are
part of a larger collection of benchmarks that are run periodically, for
which the results are collected to Grafana. We have limited resources
there, so we don't want to overload it with benchmarks for every single
control type. However, for our own benchmarking purposes it is useful to
be able to easily benchmark all of them.

Change-Id: I920e4c07bce8356e5d9f41d4c1d6ae43e37bb13e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>