Fix ASAN stack-use-after-scope error in CmpInstanceOf instruction
This is triggered by tst_qqmlecmascript::sequenceConversionThreads().
The call to fromReturnedValue(acc) creates a temporary value and the
as() cast call returns a pointer to the temp. That becomes a dangling
pointer when the temp goes out of scope.

Duplicating the fromReturnedValue() avoids this and at least gcc is
clever enough to collapse the isObject() code for both uses.

Change-Id: I741206b0c10d16fcc4ffdf68532a721f74c1b0dc
Reviewed-by: Lars Knoll <lars.knoll@qt.io>