QQuickShapePrivate: use header init, order storage by type
Header initialization makes it more clear that initialization is being
done properly (and in fact a couple of these were not).

Ordering by type allows compilers to pack the bools and enums into a
bitfield.  (Unfortunately explicit bitfields aren't syntactically
compatible with header initialization, so we leave it up to the
compiler to make this optimization on its own, since Shape is not
intended to be used in large numbers anyway.)

Change-Id: I68b1f5dec59bc2c487ab37bd3495eca19736bc21
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>