Add QUICK_TEST_MAIN_WITH_SETUP to allow executing C++ before a QML test
This macro is the same as QUICK_TEST_MAIN, but takes the user's
QObject subclass as an argument, and calls pre-defined slots/invokable
functions on it, similar to how e.g. init() is called for C++ tests.

This allows e.g. context properties to be set for the QML tests.

By basing the API on invokable functions, we give ourselves the freedom
to easily add more functions in the future.

[ChangeLog][QtQuickTest] Added QUICK_TEST_MAIN_WITH_SETUP macro to
allow executing C++ before a QML test (such as registering context
properties).

Task-number: QTBUG-50064
Change-Id: Id566e388553811c220871248403d32545f8ae1eb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>