Use shorter ':/' prefix for Imagine style example config paths
Using either :/ or qrc:/ will work because we handle both cases
in QQuickImagineStyle::url(), but for local file paths, we don't
have any handling for the file:// prefix. This patch just makes
the examples more consistent. A follow up patch will document examples
of specifying the different types of paths.

Change-Id: Ib7fc2fe012242f9b225b4380ca066107b77d0680
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>