Change default TapHandler.gesturePolicy to DragThreshold
This is based on the idea that TapHandler may be more often used to
modify existing behavior rather than building Button controls from
scratch.  DragThreshold is reasonable newbie-friendly default behavior
for both use cases.  The drag-off-drag-back-release-and-click behavior
is more advanced, and the designers of the best-behaving Button
controls can be expected to discover the need to change gesturePolicy
to get it.

Change-Id: If220acf080e04f664d020d5e832f8d16a16b857a
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>