Match QQmlEngineDebugServiceTest to the service
The fact that the service transmits QObject properties by their
objectNames, rather than as proper references is not very useful and in
fact the reason for some strange behavior in QtCreator. However, it's
been that way for a long time and we should test what the service
actually does.

Task-number: QTBUG-65852
Change-Id: I6e489bb9877fae129432b5b3d5d4c7bfb5deaa9b
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>